Merged and fixed the branches and some minor xml errors

+++ Folquest/app/src/main/java/com/kantele/folquest/Quest.java
@@ -1,7 +1,5 @@
 package com.kantele.folquest;
 
-import android.content.res.Resources;
-
 /**
  * Created by Teemu on 22.3.2017.
  */
@@ -22,19 +20,19 @@ public class Quest{
     }
 
     //Difficulty level 0-5         NOVICE,   EASY, NORMAL,   HARD, V.HARD,   EPIC
-    long pushUpDifficulty[]   =  {      5,     15,     30,     60,    200,    500 }; //reps
-    long sitUpDifficulty[]    =  {     10,     25,     50,    100,    500,   1000 }; //reps
-    long squatsDifficulty[]   =  {     10,     25,     50,    100,    500,   1000 }; //reps
-    long wallSitDifficulty[]  =  {     20,     60,    120,    300,   1500,   6000 }; //seconds
-    long distanceDifficulty[] =  {   2000,   5000,  10000,  20000, 100000, 200000 }; //meters
-    long stepsDifficulty[]    =  {   1500,   3000,   6000,  15000,  90000, 180000 }; //steps
-    long waterDifficulty[]    =  {      1,      8,     24,     56,    112,    224 }; //glasses of water for a day/day/3 days/week/2 weeks/month
-    long caloriesDifficulty[] =  {    500,   1500,   3000,   7000,  50000, 100000 }; //kilocalories burned
-    long restDifficulty[]     =  {      1,      3,      7,     14,     30,    180 }; //sleep at least 8 hours a day for this many days
+    int pushUpDifficulty[]   =  {      5,     15,     30,     60,    200,    500 }; //reps
+    int sitUpDifficulty[]    =  {     10,     25,     50,    100,    500,   1000 }; //reps
+    int squatsDifficulty[]   =  {     10,     25,     50,    100,    500,   1000 }; //reps
+    int wallSitDifficulty[]  =  {     20,     60,    120,    300,   1500,   6000 }; //seconds
+    int distanceDifficulty[] =  {   2000,   5000,  10000,  20000, 100000, 200000 }; //meters
+    int stepsDifficulty[]    =  {   1500,   3000,   6000,  15000,  90000, 180000 }; //steps
+    int waterDifficulty[]    =  {      1,      8,     24,     56,    112,    224 }; //glasses of water for a day/day/3 days/week/2 weeks/month
+    int caloriesDifficulty[] =  {    500,   1500,   3000,   7000,  50000, 100000 }; //kilocalories burned
+    int restDifficulty[]     =  {      1,      3,      7,     14,     30,    180 }; //sleep at least 8 hours a day for this many days
 
     //Rewards based on difficulty  NOVICE,   EASY, NORMAL,   HARD, V.HARD,   EPIC
-    long expRewards[]         =  {     50,    200,    500,   1250,   8000,  20000 };
-    long goldRewards[]        =  {     10,     50,    200,   1000,   4000,  20000 };
+    int expRewards[]         =  {     50,    200,    500,   1250,   8000,  20000 };
+    int goldRewards[]        =  {     10,     50,    200,   1000,   4000,  20000 };
 
     //Variables
     protected questType type;
